# WhatsApp-Clone-API/src/main/java/com/wa/whatsappclone/user/UserResponse.java
package com.wa.whatsappclone.user;

import lombok.*;

import java.time.LocalDateTime;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class UserResponse {

    private String id;
    private String keycloakId;
    private String firstName;
    private String lastName;
    private String email;
    private LocalDateTime lastSeen;
    private boolean isOnline;

}